feat(gamepad): adding gamepad and delete UtilsUi func unusued
